refactor(playground): tidy agent-builder UI and extract role-permissions fetch
- Move makeFetchRolePermissionsRequest out of role-impersonation-context
  and into domains/auth/services/fetch-role-permissions.ts as
  fetchRolePermissionsRequest.
- Memoize the RoleImpersonationProvider context value with useMemo
  instead of useCallback + inline object so the provider value is
  stable across renders.
- favorite-button: return null while useAuthCapabilities is loading so
  we don't briefly render a disabled sign-in state before capabilities
  resolve.
- agent-builder-root-layout: inline the small login-redirect URL
  builder; the helper was only used once.
