+++ packages/playground/src/domains/auth/context/role-impersonation-context.tsx
@@ -5,11 +5,11 @@
  * This is a UI-only override — server calls still use real admin permissions.
  */
 
-import type { MastraClient, RouteResponse } from '@mastra/client-js';
 import { useMastraClient } from '@mastra/react';
 import { useMutation } from '@tanstack/react-query';
-import { createContext, useCallback, useContext, useState  } from 'react';
-import type {ReactNode} from 'react';
+import { createContext, useContext, useMemo, useState } from 'react';
+import type { ReactNode } from 'react';
+import { fetchRolePermissionsRequest } from '../services/fetch-role-permissions';
 
 export type RoleImpersonationState = {
   /** The role currently being impersonated, or null */
@@ -26,39 +26,8 @@ export type RoleImpersonationState = {
   isSwitching: boolean;
 };
 
+// eslint-disable-next-line react-refresh/only-export-components
 export const RoleImpersonationContext = createContext<RoleImpersonationState | null>(null);
-
-type RolePermissionsResponse = RouteResponse<'GET /auth/roles/:roleId/permissions'>;
-
-/**
- * Makes a request to fetch the resolved permissions for a role.
- * Exported for testing purposes.
- *
- * @internal
- */
-export async function makeFetchRolePermissionsRequest(
-  client: MastraClient,
-  { roleId }: { roleId: string },
-): Promise<RolePermissionsResponse> {
-  const { baseUrl = '', apiPrefix, headers: clientHeaders = {} } = client.options || {};
-  const raw = (apiPrefix ?? '/api').trim();
-  const normalized = raw === '' ? '' : raw.startsWith('/') ? raw : `/${raw}`;
-  const prefix = normalized.replace(/\/+$/, '');
-
-  const response = await fetch(`${baseUrl}${prefix}/auth/roles/${encodeURIComponent(roleId)}/permissions`, {
-    credentials: 'include',
-    headers: {
-      ...clientHeaders,
-      'Content-Type': 'application/json',
-    },
-  });
-
-  if (!response.ok) {
-    throw new Error(`Failed to fetch role permissions: ${response.status}`);
-  }
-
-  return response.json();
-}
 
 type ImpersonationMutationResult = {
   role: { id: string; name: string };
@@ -72,7 +41,7 @@ export function RoleImpersonationProvider({ children }: { children: ReactNode })
 
   const mutation = useMutation<ImpersonationMutationResult, Error, { role: { id: string; name: string } }>({
     mutationFn: async ({ role }) => {
-      const data = await makeFetchRolePermissionsRequest(client, { roleId: role.id });
+      const data = await fetchRolePermissionsRequest(client, { roleId: role.id });
       return { role, permissions: data.permissions };
     },
     onSuccess: ({ role, permissions }) => {
@@ -83,35 +52,31 @@ export function RoleImpersonationProvider({ children }: { children: ReactNode })
 
   const { mutateAsync, reset, isPending } = mutation;
 
-  const startImpersonation = useCallback(
-    async (role: { id: string; name: string }) => {
+  const value = useMemo(() => {
+    const startImpersonation = async (role: { id: string; name: string }) => {
       await mutateAsync({ role });
-    },
-    [mutateAsync],
-  );
+    };
 
-  const stopImpersonation = useCallback(() => {
-    setImpersonatedRole(null);
-    setImpersonatedPermissions(null);
-    reset();
-  }, [reset]);
+    const stopImpersonation = () => {
+      setImpersonatedRole(null);
+      setImpersonatedPermissions(null);
+      reset();
+    };
 
-  return (
-    <RoleImpersonationContext.Provider
-      value={{
-        impersonatedRole,
-        impersonatedPermissions,
-        isImpersonating: impersonatedRole !== null,
-        startImpersonation,
-        stopImpersonation,
-        isSwitching: isPending,
-      }}
-    >
-      {children}
-    </RoleImpersonationContext.Provider>
-  );
+    return {
+      impersonatedRole,
+      impersonatedPermissions,
+      isImpersonating: impersonatedRole !== null,
+      startImpersonation,
+      stopImpersonation,
+      isSwitching: isPending,
+    };
+  }, [impersonatedRole, impersonatedPermissions, isPending, mutateAsync, reset]);
+
+  return <RoleImpersonationContext.Provider value={value}>{children}</RoleImpersonationContext.Provider>;
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
 export function useRoleImpersonation(): RoleImpersonationState {
   const ctx = useContext(RoleImpersonationContext);
   if (!ctx) {
@@ -125,5 +90,6 @@ export function useRoleImpersonation(): RoleImpersonationState {
       isSwitching: false,
     };
   }
+
   return ctx;
 }
